combine functions in one file
